WIP: Support linking packages
When working across repositories doing coordiated changes or your
local machine, it is helpful to be able to revert back to the 'link
plan', where we use yarn link to reference the other `@pulumi/*`
packages instead of restoring them from npmjs.

This change adds a sketch of what support for this would look
like. With this change, if PULUMI_LINK_PACKAGES is set in your
environment (the value does not matter), then we will run yarn link
for every `@pulumi/*` package listed in the dependencies section of
the project.json we are building or installing.

Doing so requires `jq` on the path, which may not be present on your
machine by default, but is packaged by all the major linux
distros (e.g. apt-get install jq, dnf install jq) and is also present
in homebrew (brew install jq). This is a fairly useful tool in a bunch
of other cases, and we've leaned on it in other scripts already.

This behavior is not the default. Opting into this does mean that you
may end up in gnarly cases where your local version of a dependency
has lagged behind upstream and the project you're working has taken a
dependency on that. In such cases, you'll want to either get newer
versions of dependencies and rebuild them or blow away your
node_modules folder and just used the versions from NPM.